exploratory-data-analysis---online-shopping-in-retail376/requirements.txt
pandas
matplotlib
seaborn
numpy
pyyaml
scipy
sqlalchemy
